backend_agente/requirements.txt
fastapi
uvicorn
openai
llama-index
pandas
python-dotenv
